feat: refactor authentication logic and implement user retrieval/registration in authService
